update to lynx2.8.7rel.2, with local patches: - restore local lynx.cfg settings [avsm] - fix makefile races [espie] - read/write result checking fixes to avoid unsigned comparisons vs -1 [krw] - initialize all the InputFieldData members correctly [fgsch] - fix socklen_t test to include <sys/types.h> [miod] - fgets(3) returns NULL on error, not 0. No functional change [cloder]
ok krw@, tests by Simon Kuhnle and Martin Pieuchot

+++ gnu/usr.bin/lynx/WWW/Library/Implementation/HTAssoc.c
@@ -1,5 +1,7 @@
-
-/* MODULE							HTAssoc.c
+/*
+ * $LynxId: HTAssoc.c,v 1.9 2009/01/03 01:59:16 tom Exp $
+ *
+ * MODULE							HTAssoc.c
  *	    ASSOCIATION LIST FOR STORING NAME-VALUE PAIRS.
  *	    NAMES NOT CASE SENSITIVE, AND ONLY COMMON LENGTH
  *	    IS CHECKED (allows abbreviations; well, length is
@@ -73,7 +75,7 @@ char *HTAssocList_lookup(HTAssocList *alist,
     HTAssoc *assoc;
 
     while (NULL != (assoc = (HTAssoc *) HTList_nextObject(cur))) {
-	if (!strncasecomp(assoc->name, name, strlen(name)))
+	if (!strncasecomp(assoc->name, name, (int) strlen(name)))
 	    return assoc->value;
     }
     return NULL;
